backend/requirements.txt
Django >= 4.1.3
djangorestframework >= 3.14.0
django-cors-headers >= 4.3.1
djangorestframework-simplejwt >= 5.3.1

psycopg2 >= 2.9.9
pillow >= 10.2.0
psycopg2-binary >= 2.9.9
python-decouple >= 3.8

requests >= 2.32.3

gunicorn >= 23.0.0
